2/email-service/requirements.txt
yagmail==0.11.224
kafka-python==2.0.1